prompt_engine.py
import json
import os
import random
import logging
from datetime import datetime
from typing import Dict, List
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

# Configure logging
logging.basicConfig(
    level=logging.INFO,
    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s',
    handlers=[
        logging.FileHandler("prompt_engine.log"),
        logging.StreamHandler()
    ]
)
logger = logging.getLogger("prompt_engine")

# Constants
SUPPORTED_VERTICALS = [
    "fitness",
    "finance",
    "entrepreneurship",
    "mens_health",
    "health",
    "luxury_lifestyle"
]

class PromptEngine:
    def __init__(self, trends_file: str = "daily_trends.json"):
        self.trends_file = trends_file
        self.trends_data = self._load_trends_data()
        self.performance_data = self._load_performance_data()
        self.vertical_libraries = self._load_vertical_libraries()
        self.prompt_templates = self._load_prompt_templates()

    def _load_trends_data(self) -> List[Dict]:
        """Load trends data from JSON file"""
        try:
            with open(self.trends_file, "r") as f:
                return json.load(f)
        except (FileNotFoundError, json.JSONDecodeError) as e:
            logger.error(f"Failed to load trends data: {e}")
            return []

    def _load_performance_data(self) -> Dict:
        """Load past performance data for optimization"""
        try:
            with open("performance_feedback.json", "r") as f:
                return json.load(f)
        except (FileNotFoundError, json.JSONDecodeError):
            logger.warning("No performance data found, using default weights")
            return {}

    def _load_vertical_libraries(self) -> Dict:
        """Load vertical-specific language libraries"""
        libraries = {}
        # Vertical libraries initialization (omitted for brevity)
        # ... (full library definitions as per spec) ...
        return libraries

    def _load_prompt_templates(self) -> Dict:
        """Load prompt structure templates"""
        templates = {
            "standard": {"niche": "", "hook": "", "body": "", "cta": "", "tone": ""},
            "story_based": {"niche": "", "hook": "", "story_intro": "", "problem": "", "solution": "", "result": "", "cta": "", "tone": ""},
            "expert_advice": {"niche": "", "hook": "", "authority_claim": "", "common_mistake": "", "correction": "", "evidence": "", "cta": "", "tone": ""},
            "list_format": {"niche": "", "hook": "", "intro": "", "points": [], "conclusion": "", "cta": "", "tone": ""}
        }
        return templates

    # Additional methods (get_top_trends_by_vertical, generate_content_prompt, etc.)
    # should be implemented here according to the provided spec.

def main():
    engine = PromptEngine()
    prompts_batch = engine.generate_prompts_batch(count_per_vertical=5)
    engine.save_prompts_to_json(prompts_batch)
    for vertical, prompts in prompts_batch.items():
        print(f"=== {vertical.upper()} ({len(prompts)}) ===")
        for prompt in prompts:
            print(f"HOOK: {prompt['hook']} | CTA: {prompt['cta']}")

if __name__ == "__main__":
    main()