requirements.txt
requests==2.31.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
selenium==4.12.0
webdriver-manager==4.0.1

# API clients
praw==7.7.1
pytrends==4.9.2
asyncpraw==7.7.1

# Async support
aiohttp==3.8.5
asyncio==3.4.3

# Data processing
pandas==2.1.0
numpy==1.25.2
matplotlib==3.8.0

# Testing
pytest==7.4.2
pytest-asyncio==0.21.1

# Performance and monitoring
prometheus-client==0.17.1

# Caching
diskcache==5.6.1